bitcoin-demo-v4/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
langchain>=0.1.0
faiss-cpu>=1.7.4
anthropic>=0.8.0
sentence-transformers>=2.2.2
tqdm>=4.65.0
python-dotenv>=1.0.0 